TopComponentをリファクタリング #24
