+++ client/src/actions/authActions.js
@@ -1,4 +1,5 @@
 import request from "../config/axios";
+import axios from "axios";
 
 export const signup = values => {
   return dispatch => {
@@ -32,4 +33,29 @@ export const signin = values => {
         dispatch({ type: "SIGNIN_ERROR", error: error });
       });
   };
+};
+
+export const getOAuthUrl = () => {
+  axios
+    .get("http://localhost:3001/api/github_oauth_url")
+    .then(response => {
+      const loginWindow = openLoginWindow(response.data.url);
+      loginWindow.focus();
+    })
+    .catch(error => {
+      console.log(error);
+    });
+};
+
+const openLoginWindow = url => {
+  const width = 500;
+  const height = 600;
+  const windowX = (window.screen.width - width) / 2;
+  const windowY = (window.screen.height - height) / 2;
+  const title = "ログイン - Githubアカウント";
+  window.open(
+    url,
+    title,
+    `top=${windowY},left=${windowX},width=${width},height=${height}`
+  );
 };
